+++ bin/build_package.ts
@@ -37,10 +37,12 @@ async function main() {
     const packageData = await readPackageJSON(path.join(packageRoot, 'package.json'));
     const isStimulusBundle = '@symfony/stimulus-bundle' === packageData.name;
     const isReactOrVue = ['@symfony/ux-react', '@symfony/ux-vue'].some((name) => packageData.name.startsWith(name));
+    const isTurbo = '@symfony/ux-turbo' === packageData.name;
 
     const inputCssFile = packageData?.config?.css_source;
     const inputFiles = [
         ...globSync('src/*controller.ts'),
+        ...(isTurbo ? ['src/mercure_stream_source_element.ts'] : []),
         ...(isStimulusBundle ? ['src/loader.ts', 'src/controllers.ts'] : []),
         ...(isReactOrVue ? ['src/loader.ts', 'src/components.ts'] : []),
         ...(inputCssFile ? [inputCssFile] : []),