+++ src/Turbo/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Add a minimal layout for Turbo Frame responses, allowing `head` content like meta tags to work properly
 - Add `TurboFrame` service methods and Twig functions `turbo_is_frame_request()`/`turbo_frame_request_id()` to detect Turbo Frame requests
+- Add `<turbo-mercure-stream-source>` custom element, `turbo_stream_from()` Twig function and `<twig:Turbo:Stream:From>` Twig component
+- Deprecate `turbo_stream_listen()` Twig function, use `turbo_stream_from()` or the `<twig:Turbo:Stream:From>` Twig component instead
+- Deprecate `Symfony\UX\Turbo\Twig\TurboStreamListenRendererInterface` interface, use `Symfony\UX\Turbo\StreamSourceRendererInterface` instead
+- Deprecate `Symfony\UX\Turbo\Bridge\Mercure\TurboStreamListenRenderer` class, use `Symfony\UX\Turbo\Bridge\Mercure\MercureStreamSourceRenderer` instead
+- Deprecate `Symfony\UX\Turbo\Bridge\Mercure\TopicSet` class
+- Deprecate the `mercure-turbo-stream` Stimulus controller, use the `<turbo-mercure-stream-source>` custom element instead
 
 ## 3.0.0